renamed generic parameter

+++ src/main/java/Permutation.java
@@ -5,7 +5,7 @@ public class Permutation {
     public static void main(String[] args){
 
         RandomizedQueue<String> queue = new RandomizedQueue<>();
-        int K = (int) Integer.parseInt(args[0]);
+        int K = Integer.parseInt(args[0]);
 
         try {
             do{
